chore: moving remaining tests from transpiled ref tests; minor refinements
