+++ crates/algokit_utils/tests/applications/app_client/client_management.rs
@@ -1,5 +1,3 @@
-// Tests for Client Management Features
-
 use crate::common::{AlgorandFixtureResult, TestResult, algorand_fixture, deploy_arc56_contract};
 use algokit_abi::{ABIValue, Arc56Contract};
 use algokit_transact::{OnApplicationComplete, StateSchema};
