+++ crates/algokit_utils/tests/applications/app_client/state.rs
@@ -1,5 +1,3 @@
-// AppClient state access tests: global/local state, boxes, and box maps
-
 use crate::common::{AlgorandFixtureResult, TestResult, algorand_fixture, deploy_arc56_contract};
 use algokit_abi::{ABIType, ABIValue, Arc56Contract};
 use algokit_transact::BoxReference;
@@ -13,6 +11,8 @@ use algokit_utils::{AlgorandClient as RootAlgorandClient, AppMethodCallArg, Reso
 use base64::{Engine, engine::general_purpose::STANDARD as Base64};
 use num_bigint::BigUint;
 use rstest::*;
+use std::collections::HashMap;
+use std::str::FromStr;
 use std::sync::Arc;
 
 fn get_testing_app_spec() -> Arc56Contract {
@@ -247,15 +247,16 @@ async fn test_box_retrieval(#[future] algorand_fixture: AlgorandFixtureResult) -
     let fixture = algorand_fixture.await?;
     let sender = fixture.test_account.account().address();
 
-    let mut tmpl: TealTemplateParams = Default::default();
-    tmpl.insert("VALUE".to_string(), TealTemplateValue::Int(0));
-    tmpl.insert("UPDATABLE".to_string(), TealTemplateValue::Int(0));
-    tmpl.insert("DELETABLE".to_string(), TealTemplateValue::Int(0));
     let app_id = deploy_arc56_contract(
         &fixture,
         &sender,
         &get_testing_app_spec(),
-        Some(tmpl),
+        Some(
+            [("VALUE", 0), ("UPDATABLE", 0), ("DELETABLE", 0)]
+                .into_iter()
+                .map(|(k, v)| (k.to_string(), TealTemplateValue::Int(v)))
+                .collect(),
+        ),
         None,
         None,
     )
@@ -495,6 +496,264 @@ async fn test_box_maps(#[future] algorand_fixture: AlgorandFixtureResult) -> Tes
         .get_map_value("bMap", &ABIValue::Uint(BigUint::from(1u64)))
         .await?;
     assert_eq!(box_map_value, Some(expected_value));
+
+    Ok(())
+}
+
+#[rstest]
+#[tokio::test]
+async fn box_methods_with_manually_encoded_abi_args(
+    #[future] algorand_fixture: AlgorandFixtureResult,
+) -> TestResult {
+    let fixture = algorand_fixture.await?;
+    let sender = fixture.test_account.account().address();
+
+    let json = algokit_test_artifacts::testing_app_puya::APPLICATION_ARC56;
+    let spec = Arc56Contract::from_json(json).expect("valid arc56");
+    let app_id = deploy_arc56_contract(&fixture, &sender, &spec, None, None, None).await?;
+
+    let mut algorand = RootAlgorandClient::default_localnet(None);
+    algorand.set_signer(sender.clone(), Arc::new(fixture.test_account.clone()));
+    let client = AppClient::new(AppClientParams {
+        app_id,
+        app_spec: spec,
+        algorand,
+        app_name: None,
+        default_sender: Some(sender.to_string()),
+        default_signer: None,
+        source_maps: None,
+        transaction_composer_config: None,
+    });
+
+    client
+        .fund_app_account(
+            FundAppAccountParams {
+                amount: 1_000_000,
+                sender: Some(sender.to_string()),
+                ..Default::default()
+            },
+            None,
+        )
+        .await?;
+
+    let box_prefix = b"box_bytes".to_vec();
+    let name_type = ABIType::String;
+    let box_name = "asdf";
+    let box_name_encoded = name_type.encode(&ABIValue::from(box_name)).unwrap();
+    let box_identifier = {
+        let mut v = box_prefix.clone();
+        v.extend_from_slice(&box_name_encoded);
+        v
+    };
+
+    let value_type = ABIType::DynamicArray(Box::new(ABIType::Byte));
+    let encoded = value_type
+        .encode(&ABIValue::from(vec![
+            ABIValue::from_byte(116),
+            ABIValue::from_byte(101),
+            ABIValue::from_byte(115),
+            ABIValue::from_byte(116),
+            ABIValue::from_byte(95),
+            ABIValue::from_byte(98),
+            ABIValue::from_byte(121),
+            ABIValue::from_byte(116),
+            ABIValue::from_byte(101),
+            ABIValue::from_byte(115),
+        ]))
+        .unwrap();
+
+    client
+        .send()
+        .call(
+            AppClientMethodCallParams {
+                method: "set_box_bytes".to_string(),
+                args: vec![
+                    AppMethodCallArg::ABIValue(ABIValue::from("asdf")),
+                    AppMethodCallArg::ABIValue(ABIValue::Array(
+                        encoded.into_iter().map(ABIValue::from_byte).collect(),
+                    )),
+                ],
+                sender: Some(sender.to_string()),
+                box_references: Some(vec![BoxReference {
+                    app_id: 0,
+                    name: box_identifier.clone(),
+                }]),
+                ..Default::default()
+            },
+            None,
+            None,
+        )
+        .await?;
+
+    let retrieved = client
+        .algorand()
+        .app()
+        .get_box_value_from_abi_type(client.app_id(), &box_identifier, &value_type)
+        .await?;
+    assert_eq!(
+        retrieved,
+        ABIValue::Array(vec![
+            ABIValue::from_byte(116),
+            ABIValue::from_byte(101),
+            ABIValue::from_byte(115),
+            ABIValue::from_byte(116),
+            ABIValue::from_byte(95),
+            ABIValue::from_byte(98),
+            ABIValue::from_byte(121),
+            ABIValue::from_byte(116),
+            ABIValue::from_byte(101),
+            ABIValue::from_byte(115),
+        ])
+    );
+
+    Ok(())
+}
+
+#[rstest]
+#[tokio::test]
+async fn box_methods_with_arc4_returns_parametrized(
+    #[future] algorand_fixture: AlgorandFixtureResult,
+) -> TestResult {
+    let fixture = algorand_fixture.await?;
+    let sender = fixture.test_account.account().address();
+
+    let spec =
+        Arc56Contract::from_json(algokit_test_artifacts::testing_app_puya::APPLICATION_ARC56)
+            .expect("valid arc56");
+    let app_id = deploy_arc56_contract(&fixture, &sender, &spec, None, None, None).await?;
+
+    let mut algorand = RootAlgorandClient::default_localnet(None);
+    algorand.set_signer(sender.clone(), Arc::new(fixture.test_account.clone()));
+    let client = AppClient::new(AppClientParams {
+        app_id,
+        app_spec: spec,
+        algorand,
+        app_name: None,
+        default_sender: Some(sender.to_string()),
+        default_signer: None,
+        source_maps: None,
+        transaction_composer_config: None,
+    });
+
+    client
+        .fund_app_account(
+            FundAppAccountParams {
+                amount: 1_000_000,
+                sender: Some(sender.to_string()),
+                ..Default::default()
+            },
+            None,
+        )
+        .await?;
+
+    let mut big = num_bigint::BigUint::from(1u64);
+    big <<= 256u32;
+    let cases: Vec<(Vec<u8>, &str, &str, ABIValue)> = vec![
+        (
+            b"box_str".to_vec(),
+            "set_box_str",
+            "string",
+            ABIValue::from("string"),
+        ),
+        (
+            b"box_int".to_vec(),
+            "set_box_int",
+            "uint32",
+            ABIValue::from(123u32),
+        ),
+        (
+            b"box_int512".to_vec(),
+            "set_box_int512",
+            "uint512",
+            ABIValue::from(big),
+        ),
+        (
+            b"box_static".to_vec(),
+            "set_box_static",
+            "byte[4]",
+            ABIValue::Array(vec![
+                ABIValue::from_byte(1),
+                ABIValue::from_byte(2),
+                ABIValue::from_byte(3),
+                ABIValue::from_byte(4),
+            ]),
+        ),
+        (
+            b"".to_vec(),
+            "set_struct",
+            "(string,uint64)",
+            algokit_abi::ABIValue::Array(vec![
+                algokit_abi::ABIValue::from("box1"),
+                algokit_abi::ABIValue::from(123u64),
+            ]),
+        ),
+    ];
+
+    for (box_prefix, method_sig, value_type_str, arg_val) in cases {
+        let name_type = ABIType::String;
+        let name_encoded = name_type.encode(&ABIValue::from("box1")).unwrap();
+        let mut box_reference = box_prefix.clone();
+        box_reference.extend_from_slice(&name_encoded);
+
+        let method_arg_val = if method_sig == "set_struct" {
+            algokit_abi::ABIValue::Struct(HashMap::from([
+                ("name".to_string(), algokit_abi::ABIValue::from("box1")),
+                ("id".to_string(), algokit_abi::ABIValue::from(123u64)),
+            ]))
+        } else {
+            arg_val.clone()
+        };
+
+        client
+            .send()
+            .call(
+                AppClientMethodCallParams {
+                    method: method_sig.to_string(),
+                    args: vec![
+                        AppMethodCallArg::ABIValue(ABIValue::from("box1")),
+                        AppMethodCallArg::ABIValue(method_arg_val),
+                    ],
+                    sender: Some(sender.to_string()),
+                    box_references: Some(vec![BoxReference {
+                        app_id: 0,
+                        name: box_reference.clone(),
+                    }]),
+                    ..Default::default()
+                },
+                None,
+                None,
+            )
+            .await?;
+
+        let expected_raw = algokit_abi::ABIType::from_str(value_type_str)
+            .unwrap()
+            .encode(&arg_val)
+            .unwrap();
+        let actual_raw = client.get_box_value(&box_reference).await?;
+        assert_eq!(actual_raw, expected_raw);
+
+        let decoded = client
+            .get_box_value_from_abi_type(
+                &box_reference,
+                &ABIType::from_str(value_type_str).unwrap(),
+            )
+            .await?;
+        assert_eq!(decoded, arg_val);
+
+        if method_sig == "set_struct" {
+            let struct_box_name = box_reference.clone();
+            let values = client
+                .get_box_values_from_abi_type(
+                    &ABIType::from_str(value_type_str).unwrap(),
+                    Some(Box::new(move |name: &BoxName| {
+                        name.name_raw == struct_box_name
+                    })),
+                )
+                .await?;
+            assert_eq!(values.len(), 1);
+            assert_eq!(values[0].value, decoded);
+        }
+    }
 
     Ok(())
 }
