--- crates/algokit_utils/tests/applications/ref.rs
@@ -1,1362 +1,0 @@
-use crate::common::{AlgorandFixtureResult, TestResult, algorand_fixture, deploy_arc56_contract};
-use algokit_abi::{ABIType, ABIValue, Arc56Contract};
-use algokit_transact::BoxReference;
-use algokit_utils::applications::app_client::{AppClient, AppClientParams};
-use algokit_utils::applications::app_client::{AppClientMethodCallParams, FundAppAccountParams};
-use algokit_utils::clients::app_manager::{TealTemplateParams, TealTemplateValue};
-use algokit_utils::{
-    AlgorandClient as RootAlgorandClient, AppMethodCallArg, AppSourceMaps, PaymentParams,
-    TransactionResultError,
-};
-use rstest::*;
-use std::str::FromStr;
-use std::sync::Arc;
-
-fn get_testing_app_spec() -> Arc56Contract {
-    let json = algokit_test_artifacts::testing_app::APPLICATION_ARC56;
-    Arc56Contract::from_json(json).expect("valid arc56")
-}
-
-fn get_sandbox_spec() -> Arc56Contract {
-    let json = algokit_test_artifacts::sandbox::APPLICATION_ARC56;
-    Arc56Contract::from_json(json).expect("valid arc56")
-}
-
-#[rstest]
-#[tokio::test]
-async fn retrieve_state(#[future] algorand_fixture: AlgorandFixtureResult) -> TestResult {
-    let fixture = algorand_fixture.await?;
-    let sender = fixture.test_account.account().address();
-
-    // Deploy testing_app
-    let mut tmpl: TealTemplateParams = Default::default();
-    tmpl.insert("VALUE".to_string(), TealTemplateValue::Int(0));
-    tmpl.insert("UPDATABLE".to_string(), TealTemplateValue::Int(0));
-    tmpl.insert("DELETABLE".to_string(), TealTemplateValue::Int(0));
-    let app_id =
-        deploy_arc56_contract(&fixture, &sender, &get_testing_app_spec(), Some(tmpl), None).await?;
-
-    let mut algorand = RootAlgorandClient::default_localnet(None);
-    algorand.set_signer(sender.clone(), Arc::new(fixture.test_account.clone()));
-    let client = AppClient::new(AppClientParams {
-        app_id,
-        app_spec: get_testing_app_spec(),
-        algorand,
-        app_name: None,
-        default_sender: Some(sender.to_string()),
-        default_signer: None,
-        source_maps: None,
-        transaction_composer_config: None,
-    });
-
-    // Global state: set and verify
-    client
-        .send()
-        .call(
-            AppClientMethodCallParams {
-                method: "set_global".to_string(),
-                args: vec![
-                    AppMethodCallArg::ABIValue(ABIValue::from(1u64)),
-                    AppMethodCallArg::ABIValue(ABIValue::from(2u64)),
-                    AppMethodCallArg::ABIValue(ABIValue::from("asdf")),
-                    AppMethodCallArg::ABIValue(ABIValue::Array(vec![
-                        ABIValue::from_byte(1),
-                        ABIValue::from_byte(2),
-                        ABIValue::from_byte(3),
-                        ABIValue::from_byte(4),
-                    ])),
-                ],
-                sender: Some(sender.to_string()),
-                ..Default::default()
-            },
-            None,
-            None,
-        )
-        .await?;
-
-    let global_state = client.state().global_state().get_all().await?;
-    assert!(global_state.contains_key("int1"));
-    assert!(global_state.contains_key("int2"));
-    assert!(global_state.contains_key("bytes1"));
-    assert!(global_state.contains_key("bytes2"));
-    assert_eq!(
-        global_state.get("int1").unwrap().as_ref().unwrap(),
-        &ABIValue::from(1u64)
-    );
-    assert_eq!(
-        global_state.get("int2").unwrap().as_ref().unwrap(),
-        &ABIValue::from(2u64)
-    );
-    assert_eq!(
-        global_state.get("bytes1").unwrap().as_ref().unwrap(),
-        &ABIValue::from("asdf")
-    );
-
-    // Local: opt-in and set; verify
-    client
-        .send()
-        .opt_in(
-            AppClientMethodCallParams {
-                method: "opt_in".to_string(),
-                args: vec![],
-                sender: Some(sender.to_string()),
-                rekey_to: None,
-                note: None,
-                lease: None,
-                static_fee: None,
-                extra_fee: None,
-                max_fee: None,
-                validity_window: None,
-                first_valid_round: None,
-                last_valid_round: None,
-                account_references: None,
-                app_references: None,
-                asset_references: None,
-                box_references: None,
-            },
-            None,
-        )
-        .await?;
-
-    client
-        .send()
-        .call(
-            AppClientMethodCallParams {
-                method: "set_local".to_string(),
-                args: vec![
-                    AppMethodCallArg::ABIValue(ABIValue::from(1u64)),
-                    AppMethodCallArg::ABIValue(ABIValue::from(2u64)),
-                    AppMethodCallArg::ABIValue(ABIValue::from("asdf")),
-                    AppMethodCallArg::ABIValue(ABIValue::Array(vec![
-                        ABIValue::from_byte(1),
-                        ABIValue::from_byte(2),
-                        ABIValue::from_byte(3),
-                        ABIValue::from_byte(4),
-                    ])),
-                ],
-                sender: Some(sender.to_string()),
-                rekey_to: None,
-                note: None,
-                lease: None,
-                static_fee: None,
-                extra_fee: None,
-                max_fee: None,
-                validity_window: None,
-                first_valid_round: None,
-                last_valid_round: None,
-                account_references: None,
-                app_references: None,
-                asset_references: None,
-                box_references: None,
-            },
-            None,
-            None,
-        )
-        .await?;
-
-    let local_state = client
-        .state()
-        .local_state(&sender.to_string())
-        .get_all()
-        .await?;
-    assert_eq!(
-        local_state.get("local_int1").unwrap().as_ref().unwrap(),
-        &ABIValue::from(1u64)
-    );
-    assert_eq!(
-        local_state.get("local_int2").unwrap().as_ref().unwrap(),
-        &ABIValue::from(2u64)
-    );
-    assert_eq!(
-        local_state.get("local_bytes1").unwrap().as_ref().unwrap(),
-        &ABIValue::from("asdf")
-    );
-
-    // Boxes
-    let box_name1: Vec<u8> = vec![0, 0, 0, 1];
-    let box_name2: Vec<u8> = vec![0, 0, 0, 2];
-
-    // Fund app account to enable box writes
-    client
-        .fund_app_account(
-            FundAppAccountParams {
-                amount: 1_000_000,
-                sender: Some(sender.to_string()),
-                ..Default::default()
-            },
-            None,
-        )
-        .await?;
-
-    client
-        .send()
-        .call(
-            AppClientMethodCallParams {
-                method: "set_box".to_string(),
-                args: vec![
-                    AppMethodCallArg::ABIValue(ABIValue::Array(
-                        box_name1.iter().copied().map(ABIValue::from_byte).collect(),
-                    )),
-                    AppMethodCallArg::ABIValue(ABIValue::from("value1")),
-                ],
-                sender: Some(sender.to_string()),
-                rekey_to: None,
-                note: None,
-                lease: None,
-                static_fee: None,
-                extra_fee: None,
-                max_fee: None,
-                validity_window: None,
-                first_valid_round: None,
-                last_valid_round: None,
-                account_references: None,
-                app_references: None,
-                asset_references: None,
-                box_references: Some(vec![BoxReference {
-                    app_id: 0,
-                    name: box_name1.clone(),
-                }]),
-            },
-            None,
-            None,
-        )
-        .await?;
-
-    client
-        .send()
-        .call(
-            AppClientMethodCallParams {
-                method: "set_box".to_string(),
-                args: vec![
-                    AppMethodCallArg::ABIValue(ABIValue::Array(
-                        box_name2.iter().copied().map(ABIValue::from_byte).collect(),
-                    )),
-                    AppMethodCallArg::ABIValue(ABIValue::from("value2")),
-                ],
-                sender: Some(sender.to_string()),
-                rekey_to: None,
-                note: None,
-                lease: None,
-                static_fee: None,
-                extra_fee: None,
-                max_fee: None,
-                validity_window: None,
-                first_valid_round: None,
-                last_valid_round: None,
-                account_references: None,
-                app_references: None,
-                asset_references: None,
-                box_references: Some(vec![BoxReference {
-                    app_id: 0,
-                    name: box_name2.clone(),
-                }]),
-            },
-            None,
-            None,
-        )
-        .await?;
-
-    let box_names = client.get_box_names().await?;
-    let names: Vec<Vec<u8>> = box_names.into_iter().map(|n| n.name_raw).collect();
-    assert!(names.contains(&box_name1));
-    assert!(names.contains(&box_name2));
-
-    let box1_value = client.get_box_value(&box_name1).await?;
-    assert_eq!(box1_value, b"value1");
-
-    Ok(())
-}
-
-#[rstest]
-#[tokio::test]
-async fn logic_error_exposure_with_source_maps(
-    #[future] algorand_fixture: AlgorandFixtureResult,
-) -> TestResult {
-    let fixture = algorand_fixture.await?;
-    let sender = fixture.test_account.account().address();
-
-    // Deploy testing_app with template params
-    let mut tmpl: TealTemplateParams = Default::default();
-    tmpl.insert("VALUE".to_string(), TealTemplateValue::Int(0));
-    tmpl.insert("UPDATABLE".to_string(), TealTemplateValue::Int(0));
-    tmpl.insert("DELETABLE".to_string(), TealTemplateValue::Int(0));
-    let app_id = deploy_arc56_contract(
-        &fixture,
-        &sender,
-        &get_testing_app_spec(),
-        Some(tmpl.clone()),
-        None,
-        None,
-        None,
-    )
-    .await?;
-
-    let mut algorand = RootAlgorandClient::default_localnet(None);
-    algorand.set_signer(sender.clone(), Arc::new(fixture.test_account.clone()));
-    let mut client = AppClient::new(AppClientParams {
-        app_id,
-        app_spec: get_testing_app_spec(),
-        algorand,
-        app_name: None,
-        default_sender: Some(sender.to_string()),
-        default_signer: None,
-        source_maps: None,
-        transaction_composer_config: None,
-    });
-
-    // Compile TEAL to get source maps and import
-    let src = client.app_spec().source.as_ref().expect("source expected");
-    let approval_teal = src.get_decoded_approval().unwrap();
-    let clear_teal = src.get_decoded_clear().unwrap();
-    let app_manager = fixture.algorand_client.app();
-    let compiled_approval = app_manager
-        .compile_teal_template(&approval_teal, Some(&tmpl), None)
-        .await?;
-    let compiled_clear = app_manager
-        .compile_teal_template(&clear_teal, Some(&tmpl), None)
-        .await?;
-    client.import_source_maps(AppSourceMaps {
-        approval_source_map: compiled_approval.source_map,
-        clear_source_map: compiled_clear.source_map,
-    });
-
-    // Trigger logic error
-    let err = client
-        .send()
-        .call(
-            AppClientMethodCallParams {
-                method: "error".to_string(),
-                args: vec![],
-                sender: Some(sender.to_string()),
-                rekey_to: None,
-                note: None,
-                lease: None,
-                static_fee: None,
-                extra_fee: None,
-                max_fee: None,
-                validity_window: None,
-                first_valid_round: None,
-                last_valid_round: None,
-                account_references: None,
-                app_references: None,
-                asset_references: None,
-                box_references: None,
-            },
-            None,
-            None,
-        )
-        .await
-        .expect_err("expected error");
-
-    let logic = client.expose_logic_error(
-        &TransactionResultError::ParsingError {
-            message: err.to_string(),
-        },
-        false,
-    );
-    assert!(logic.pc.is_some());
-    assert!(logic.logic_error_str.contains("assert failed"));
-    if let Some(id) = &logic.transaction_id {
-        assert!(id.len() >= 52);
-    }
-    Ok(())
-}
-
-#[rstest]
-#[tokio::test]
-async fn box_methods_with_manually_encoded_abi_args(
-    #[future] algorand_fixture: AlgorandFixtureResult,
-) -> TestResult {
-    let fixture = algorand_fixture.await?;
-    let sender = fixture.test_account.account().address();
-
-    // Deploy testing_app_puya
-    let json = algokit_test_artifacts::testing_app_puya::APPLICATION_ARC56;
-    let spec = Arc56Contract::from_json(json).expect("valid arc56");
-    let app_id = deploy_arc56_contract(&fixture, &sender, &spec, None, None).await?;
-
-    let mut algorand = RootAlgorandClient::default_localnet(None);
-    algorand.set_signer(sender.clone(), Arc::new(fixture.test_account.clone()));
-    let client = AppClient::new(AppClientParams {
-        app_id,
-        app_spec: spec,
-        algorand,
-        app_name: None,
-        default_sender: Some(sender.to_string()),
-        default_signer: None,
-        source_maps: None,
-        transaction_composer_config: None,
-    });
-
-    // Fund app account
-    client
-        .fund_app_account(
-            FundAppAccountParams {
-                amount: 1_000_000,
-                sender: Some(sender.to_string()),
-                ..Default::default()
-            },
-            None,
-        )
-        .await?;
-
-    // Prepare box name and encoded value
-    let box_prefix = b"box_bytes".to_vec();
-    let name_type = ABIType::from_str("string").unwrap();
-    let box_name = "asdf";
-    let box_name_encoded = name_type.encode(&ABIValue::from(box_name)).unwrap();
-    let box_identifier = {
-        let mut v = box_prefix.clone();
-        v.extend_from_slice(&box_name_encoded);
-        v
-    };
-
-    // byte[] value
-    let value_type = ABIType::from_str("byte[]").unwrap();
-    let encoded = value_type
-        .encode(&ABIValue::from(vec![
-            ABIValue::from_byte(116),
-            ABIValue::from_byte(101),
-            ABIValue::from_byte(115),
-            ABIValue::from_byte(116),
-            ABIValue::from_byte(95),
-            ABIValue::from_byte(98),
-            ABIValue::from_byte(121),
-            ABIValue::from_byte(116),
-            ABIValue::from_byte(101),
-            ABIValue::from_byte(115),
-        ]))
-        .unwrap();
-
-    client
-        .send()
-        .call(
-            AppClientMethodCallParams {
-                method: "set_box_bytes".to_string(),
-                args: vec![
-                    AppMethodCallArg::ABIValue(ABIValue::from("asdf")),
-                    AppMethodCallArg::ABIValue(ABIValue::Array(
-                        encoded.into_iter().map(ABIValue::from_byte).collect(),
-                    )),
-                ],
-                sender: Some(sender.to_string()),
-                rekey_to: None,
-                note: None,
-                lease: None,
-                static_fee: None,
-                extra_fee: None,
-                max_fee: None,
-                validity_window: None,
-                first_valid_round: None,
-                last_valid_round: None,
-                account_references: None,
-                app_references: None,
-                asset_references: None,
-                box_references: Some(vec![BoxReference {
-                    app_id: 0,
-                    name: box_identifier.clone(),
-                }]),
-            },
-            None,
-            None,
-        )
-        .await?;
-
-    let retrieved = client
-        .algorand()
-        .app()
-        .get_box_value_from_abi_type(client.app_id(), &box_identifier, &value_type)
-        .await?;
-    assert_eq!(
-        retrieved,
-        ABIValue::Array(vec![
-            ABIValue::from_byte(116),
-            ABIValue::from_byte(101),
-            ABIValue::from_byte(115),
-            ABIValue::from_byte(116),
-            ABIValue::from_byte(95),
-            ABIValue::from_byte(98),
-            ABIValue::from_byte(121),
-            ABIValue::from_byte(116),
-            ABIValue::from_byte(101),
-            ABIValue::from_byte(115),
-        ])
-    );
-
-    Ok(())
-}
-
-#[rstest]
-#[tokio::test]
-async fn construct_transaction_with_abi_encoding_including_foreign_references_not_in_signature(
-    #[future] algorand_fixture: AlgorandFixtureResult,
-) -> TestResult {
-    let fixture = algorand_fixture.await?;
-    let sender = fixture.test_account.account().address();
-    // Deploy testing_app which has call_abi_foreign_refs()string
-    let mut tmpl: TealTemplateParams = Default::default();
-    tmpl.insert("VALUE".to_string(), TealTemplateValue::Int(0));
-    tmpl.insert("UPDATABLE".to_string(), TealTemplateValue::Int(0));
-    tmpl.insert("DELETABLE".to_string(), TealTemplateValue::Int(0));
-    let app_id =
-        deploy_arc56_contract(&fixture, &sender, &get_testing_app_spec(), Some(tmpl), None).await?;
-
-    let mut algorand = RootAlgorandClient::default_localnet(None);
-    algorand.set_signer(sender.clone(), Arc::new(fixture.test_account.clone()));
-
-    let client = AppClient::new(AppClientParams {
-        app_id,
-        app_spec: get_testing_app_spec(),
-        algorand,
-        app_name: None,
-        default_sender: Some(sender.to_string()),
-        default_signer: None,
-        source_maps: None,
-        transaction_composer_config: None,
-    });
-
-    // Create a secondary account for account_references
-    let mut new_fixture = fixture; // reuse underlying clients for funding convenience
-    let new_account = new_fixture.generate_account(None).await?;
-    let new_addr = new_account.account().address();
-
-    let send_res = client
-        .send()
-        .call(
-            AppClientMethodCallParams {
-                method: "call_abi_foreign_refs".to_string(),
-                args: vec![],
-                sender: Some(sender.to_string()),
-                rekey_to: None,
-                note: None,
-                lease: None,
-                static_fee: None,
-                extra_fee: None,
-                max_fee: None,
-                validity_window: None,
-                first_valid_round: None,
-                last_valid_round: None,
-                account_references: Some(vec![new_addr.to_string()]),
-                app_references: Some(vec![345]),
-                asset_references: Some(vec![567]),
-                box_references: None,
-            },
-            None,
-            None,
-        )
-        .await?;
-
-    let abi_ret = send_res.abi_return.as_ref().expect("abi return expected");
-    if let Some(ABIValue::String(s)) = &abi_ret.return_value {
-        assert!(s.contains("App: 345"));
-        assert!(s.contains("Asset: 567"));
-    } else {
-        panic!("expected string return");
-    }
-    Ok(())
-}
-
-#[rstest]
-#[tokio::test]
-async fn abi_with_default_arg_from_local_state(
-    #[future] algorand_fixture: AlgorandFixtureResult,
-) -> TestResult {
-    let fixture = algorand_fixture.await?;
-    let sender = fixture.test_account.account().address();
-    let mut tmpl: TealTemplateParams = Default::default();
-    tmpl.insert("VALUE".to_string(), TealTemplateValue::Int(0));
-    tmpl.insert("UPDATABLE".to_string(), TealTemplateValue::Int(0));
-    tmpl.insert("DELETABLE".to_string(), TealTemplateValue::Int(0));
-    let app_id =
-        deploy_arc56_contract(&fixture, &sender, &get_testing_app_spec(), Some(tmpl), None).await?;
-
-    let mut algorand = RootAlgorandClient::default_localnet(None);
-    algorand.set_signer(sender.clone(), Arc::new(fixture.test_account.clone()));
-
-    let client = AppClient::new(AppClientParams {
-        app_id,
-        app_spec: get_testing_app_spec(),
-        algorand,
-        app_name: None,
-        default_sender: Some(sender.to_string()),
-        default_signer: None,
-        source_maps: None,
-        transaction_composer_config: None,
-    });
-
-    // Opt-in and set local state
-
-    client
-        .send()
-        .opt_in(
-            AppClientMethodCallParams {
-                method: "opt_in".to_string(),
-                args: vec![],
-                sender: Some(sender.to_string()),
-                rekey_to: None,
-                note: None,
-                lease: None,
-                static_fee: None,
-                extra_fee: None,
-                max_fee: None,
-                validity_window: None,
-                first_valid_round: None,
-                last_valid_round: None,
-                account_references: None,
-                app_references: None,
-                asset_references: None,
-                box_references: None,
-            },
-            None,
-        )
-        .await?;
-
-    client
-        .send()
-        .call(
-            AppClientMethodCallParams {
-                method: "set_local".to_string(),
-                args: vec![
-                    AppMethodCallArg::ABIValue(ABIValue::from(1u64)),
-                    AppMethodCallArg::ABIValue(ABIValue::from(2u64)),
-                    AppMethodCallArg::ABIValue(ABIValue::from("bananas")),
-                    AppMethodCallArg::ABIValue(ABIValue::Array(vec![
-                        ABIValue::from_byte(1),
-                        ABIValue::from_byte(2),
-                        ABIValue::from_byte(3),
-                        ABIValue::from_byte(4),
-                    ])),
-                ],
-                sender: Some(sender.to_string()),
-                rekey_to: None,
-                note: None,
-                lease: None,
-                static_fee: None,
-                extra_fee: None,
-                max_fee: None,
-                validity_window: None,
-                first_valid_round: None,
-                last_valid_round: None,
-                account_references: None,
-                app_references: None,
-                asset_references: None,
-                box_references: None,
-            },
-            None,
-            None,
-        )
-        .await?;
-
-    // Call with explicit value first; expect echo prefix + defined value
-    let defined = client
-        .send()
-        .call(
-            AppClientMethodCallParams {
-                method: "default_value_from_local_state".to_string(),
-                args: vec![AppMethodCallArg::ABIValue(ABIValue::from("defined value"))],
-                sender: Some(sender.to_string()),
-                rekey_to: None,
-                note: None,
-                lease: None,
-                static_fee: None,
-                extra_fee: None,
-                max_fee: None,
-                validity_window: None,
-                first_valid_round: None,
-                last_valid_round: None,
-                account_references: None,
-                app_references: None,
-                asset_references: None,
-                box_references: None,
-            },
-            None,
-            None,
-        )
-        .await?;
-    let defined_ret = defined.abi_return.as_ref().expect("abi return expected");
-    match &defined_ret.return_value {
-        Some(ABIValue::String(s)) => {
-            assert_eq!(s, "Local state, defined value");
-        }
-        _ => panic!("expected string return"),
-    }
-
-    // Call method without providing arg; expect default from local state ("bananas")
-    let res = client
-        .send()
-        .call(
-            AppClientMethodCallParams {
-                method: "default_value_from_local_state".to_string(),
-                args: vec![AppMethodCallArg::DefaultValue],
-                sender: Some(sender.to_string()),
-                rekey_to: None,
-                note: None,
-                lease: None,
-                static_fee: None,
-                extra_fee: None,
-                max_fee: None,
-                validity_window: None,
-                first_valid_round: None,
-                last_valid_round: None,
-                account_references: None,
-                app_references: None,
-                asset_references: None,
-                box_references: None,
-            },
-            None,
-            None,
-        )
-        .await?;
-
-    let abi_ret = res.abi_return.as_ref().expect("abi return expected");
-    if let Some(ABIValue::String(s)) = &abi_ret.return_value {
-        assert_eq!(s, "Local state, bananas"); // TODO: confirm this, the current logic doesn't automatically convert base64 to utf8
-    } else {
-        panic!("expected string return");
-    }
-    Ok(())
-}
-
-#[rstest]
-#[tokio::test]
-async fn abi_with_default_arg_from_literal(
-    #[future] algorand_fixture: AlgorandFixtureResult,
-) -> TestResult {
-    // Mirrors TS: default_value(string)string
-    let fixture = algorand_fixture.await?;
-    let sender = fixture.test_account.account().address();
-
-    let mut tmpl: TealTemplateParams = Default::default();
-    tmpl.insert("VALUE".to_string(), TealTemplateValue::Int(0));
-    tmpl.insert("UPDATABLE".to_string(), TealTemplateValue::Int(0));
-    tmpl.insert("DELETABLE".to_string(), TealTemplateValue::Int(0));
-    let app_id =
-        deploy_arc56_contract(&fixture, &sender, &get_testing_app_spec(), Some(tmpl), None).await?;
-
-    let mut algorand = RootAlgorandClient::default_localnet(None);
-    algorand.set_signer(sender.clone(), Arc::new(fixture.test_account.clone()));
-    let client = AppClient::new(AppClientParams {
-        app_id,
-        app_spec: get_testing_app_spec(),
-        algorand,
-        app_name: None,
-        default_sender: Some(sender.to_string()),
-        default_signer: None,
-        source_maps: None,
-        transaction_composer_config: None,
-    });
-
-    // Call with explicit value
-    let defined = client
-        .send()
-        .call(
-            AppClientMethodCallParams {
-                method: "default_value".to_string(),
-                args: vec![AppMethodCallArg::ABIValue(ABIValue::from("defined value"))],
-                sender: Some(sender.to_string()),
-                ..Default::default()
-            },
-            None,
-            None,
-        )
-        .await?;
-    let defined_ret = defined.abi_return.as_ref().expect("abi return expected");
-    match &defined_ret.return_value {
-        Some(ABIValue::String(s)) => {
-            assert_eq!(s, "defined value");
-        }
-        _ => panic!("expected string return"),
-    }
-
-    // Call with default (no arg)
-    let defaulted = client
-        .send()
-        .call(
-            AppClientMethodCallParams {
-                method: "default_value".to_string(),
-                args: vec![AppMethodCallArg::DefaultValue],
-                sender: Some(sender.to_string()),
-                ..Default::default()
-            },
-            None,
-            None,
-        )
-        .await?;
-    let default_ret = defaulted.abi_return.as_ref().expect("abi return expected");
-    match &default_ret.return_value {
-        Some(ABIValue::String(s)) => {
-            assert_eq!(s, "default value");
-        }
-        _ => panic!("expected string return"),
-    }
-    Ok(())
-}
-
-#[rstest]
-#[tokio::test]
-async fn abi_with_default_arg_from_method(
-    #[future] algorand_fixture: AlgorandFixtureResult,
-) -> TestResult {
-    // Mirrors TS: default_value_from_abi(string)string
-    let fixture = algorand_fixture.await?;
-    let sender = fixture.test_account.account().address();
-
-    let mut tmpl: TealTemplateParams = Default::default();
-    tmpl.insert("VALUE".to_string(), TealTemplateValue::Int(0));
-    tmpl.insert("UPDATABLE".to_string(), TealTemplateValue::Int(0));
-    tmpl.insert("DELETABLE".to_string(), TealTemplateValue::Int(0));
-    let app_id =
-        deploy_arc56_contract(&fixture, &sender, &get_testing_app_spec(), Some(tmpl), None).await?;
-
-    let mut algorand = RootAlgorandClient::default_localnet(None);
-    algorand.set_signer(sender.clone(), Arc::new(fixture.test_account.clone()));
-    let client = AppClient::new(AppClientParams {
-        app_id,
-        app_spec: get_testing_app_spec(),
-        algorand,
-        app_name: None,
-        default_sender: Some(sender.to_string()),
-        default_signer: None,
-        source_maps: None,
-        transaction_composer_config: None,
-    });
-
-    // Call with explicit value
-    let defined = client
-        .send()
-        .call(
-            AppClientMethodCallParams {
-                method: "default_value_from_abi".to_string(),
-                args: vec![AppMethodCallArg::ABIValue(ABIValue::from("defined value"))],
-                sender: Some(sender.to_string()),
-                ..Default::default()
-            },
-            None,
-            None,
-        )
-        .await?;
-    let defined_ret = defined.abi_return.as_ref().expect("abi return expected");
-    match &defined_ret.return_value {
-        Some(ABIValue::String(s)) => {
-            assert_eq!(s, "ABI, defined value");
-        }
-        _ => panic!("expected string return"),
-    }
-
-    // Call with default (no arg)
-    let defaulted = client
-        .send()
-        .call(
-            AppClientMethodCallParams {
-                method: "default_value_from_abi".to_string(),
-                args: vec![AppMethodCallArg::DefaultValue],
-                sender: Some(sender.to_string()),
-                ..Default::default()
-            },
-            None,
-            None,
-        )
-        .await?;
-    let default_ret = defaulted.abi_return.as_ref().expect("abi return expected");
-    match &default_ret.return_value {
-        Some(ABIValue::String(s)) => {
-            assert_eq!(s, "ABI, default value");
-        }
-        _ => panic!("expected string return"),
-    }
-    Ok(())
-}
-
-#[rstest]
-#[tokio::test]
-async fn abi_with_default_arg_from_global_state(
-    #[future] algorand_fixture: AlgorandFixtureResult,
-) -> TestResult {
-    // Mirrors TS: default_value_from_global_state(uint64)uint64
-    let fixture = algorand_fixture.await?;
-    let sender = fixture.test_account.account().address();
-
-    let mut tmpl: TealTemplateParams = Default::default();
-    tmpl.insert("VALUE".to_string(), TealTemplateValue::Int(0));
-    tmpl.insert("UPDATABLE".to_string(), TealTemplateValue::Int(0));
-    tmpl.insert("DELETABLE".to_string(), TealTemplateValue::Int(0));
-    let app_id =
-        deploy_arc56_contract(&fixture, &sender, &get_testing_app_spec(), Some(tmpl), None).await?;
-
-    let mut algorand = RootAlgorandClient::default_localnet(None);
-    algorand.set_signer(sender.clone(), Arc::new(fixture.test_account.clone()));
-    let client = AppClient::new(AppClientParams {
-        app_id,
-        app_spec: get_testing_app_spec(),
-        algorand,
-        app_name: None,
-        default_sender: Some(sender.to_string()),
-        default_signer: None,
-        source_maps: None,
-        transaction_composer_config: None,
-    });
-
-    // Seed global state (int1) via set_global
-    let seeded_val: u64 = 456;
-    client
-        .send()
-        .call(
-            AppClientMethodCallParams {
-                method: "set_global".to_string(),
-                args: vec![
-                    AppMethodCallArg::ABIValue(ABIValue::from(seeded_val)),
-                    AppMethodCallArg::ABIValue(ABIValue::from(2u64)),
-                    AppMethodCallArg::ABIValue(ABIValue::from("asdf")),
-                    AppMethodCallArg::ABIValue(ABIValue::Array(vec![
-                        ABIValue::from_byte(1),
-                        ABIValue::from_byte(2),
-                        ABIValue::from_byte(3),
-                        ABIValue::from_byte(4),
-                    ])),
-                ],
-                sender: Some(sender.to_string()),
-                ..Default::default()
-            },
-            None,
-            None,
-        )
-        .await?;
-
-    // Call with explicit value
-    let defined = client
-        .send()
-        .call(
-            AppClientMethodCallParams {
-                method: "default_value_from_global_state".to_string(),
-                args: vec![AppMethodCallArg::ABIValue(ABIValue::from(123u64))],
-                sender: Some(sender.to_string()),
-                ..Default::default()
-            },
-            None,
-            None,
-        )
-        .await?;
-    let defined_ret = defined.abi_return.as_ref().expect("abi return expected");
-    match &defined_ret.return_value {
-        Some(ABIValue::Uint(u)) => {
-            assert_eq!(*u, num_bigint::BigUint::from(123u64));
-        }
-        _ => panic!("expected uint return"),
-    }
-
-    // Call with default (no arg) -> should read seeded global state
-    let defaulted = client
-        .send()
-        .call(
-            AppClientMethodCallParams {
-                method: "default_value_from_global_state".to_string(),
-                args: vec![AppMethodCallArg::DefaultValue],
-                sender: Some(sender.to_string()),
-                ..Default::default()
-            },
-            None,
-            None,
-        )
-        .await?;
-    let default_ret = defaulted.abi_return.as_ref().expect("abi return expected");
-    match &default_ret.return_value {
-        Some(ABIValue::Uint(u)) => {
-            assert_eq!(*u, num_bigint::BigUint::from(seeded_val));
-        }
-        _ => panic!("expected uint return"),
-    }
-    Ok(())
-}
-#[rstest]
-#[tokio::test]
-async fn bare_call_with_box_reference_builds_and_sends(
-    #[future] algorand_fixture: AlgorandFixtureResult,
-) -> TestResult {
-    let fixture = algorand_fixture.await?;
-    let sender = fixture.test_account.account().address();
-
-    let app_id = deploy_arc56_contract(&fixture, &sender, &get_sandbox_spec(), None, None).await?;
-
-    let mut algorand = RootAlgorandClient::default_localnet(None);
-    algorand.set_signer(sender.clone(), Arc::new(fixture.test_account.clone()));
-
-    let client = AppClient::new(AppClientParams {
-        app_id,
-        app_spec: get_sandbox_spec(),
-        algorand,
-        app_name: None,
-        default_sender: Some(sender.to_string()),
-        default_signer: None,
-        source_maps: None,
-        transaction_composer_config: None,
-    });
-
-    // Use method call (sandbox does not allow bare NoOp)
-
-    let result = client
-        .send()
-        .call(
-            AppClientMethodCallParams {
-                method: "hello_world".to_string(),
-                args: vec![AppMethodCallArg::ABIValue(ABIValue::from("test"))],
-                sender: Some(sender.to_string()),
-                rekey_to: None,
-                note: None,
-                lease: None,
-                static_fee: None,
-                extra_fee: None,
-                max_fee: None,
-                validity_window: None,
-                first_valid_round: None,
-                last_valid_round: None,
-                account_references: None,
-                app_references: None,
-                asset_references: None,
-                box_references: Some(vec![BoxReference {
-                    app_id: 0,
-                    name: b"1".to_vec(),
-                }]),
-            },
-            None,
-            None,
-        )
-        .await?;
-
-    match &result.common_params.transaction {
-        algokit_transact::Transaction::AppCall(fields) => {
-            assert_eq!(fields.app_id, app_id);
-            assert_eq!(
-                fields.box_references.as_ref().unwrap(),
-                &vec![BoxReference {
-                    app_id: 0,
-                    name: b"1".to_vec()
-                }]
-            );
-        }
-        _ => panic!("expected app call"),
-    }
-
-    Ok(())
-}
-
-#[rstest]
-#[tokio::test]
-async fn construct_transaction_with_boxes(
-    #[future] algorand_fixture: AlgorandFixtureResult,
-) -> TestResult {
-    let fixture = algorand_fixture.await?;
-    let sender = fixture.test_account.account().address();
-    let app_id = deploy_arc56_contract(&fixture, &sender, &get_sandbox_spec(), None, None).await?;
-
-    let mut algorand = RootAlgorandClient::default_localnet(None);
-    algorand.set_signer(sender.clone(), Arc::new(fixture.test_account.clone()));
-
-    let client = AppClient::new(AppClientParams {
-        app_id,
-        app_spec: get_sandbox_spec(),
-        algorand,
-        app_name: None,
-        default_sender: Some(sender.to_string()),
-        default_signer: None,
-        source_maps: None,
-        transaction_composer_config: None,
-    });
-
-    // Build transaction with a box reference
-
-    let built = client
-        .create_transaction()
-        .call(
-            AppClientMethodCallParams {
-                method: "hello_world".to_string(),
-                args: vec![AppMethodCallArg::ABIValue(ABIValue::from("test"))],
-                sender: Some(sender.to_string()),
-                rekey_to: None,
-                note: None,
-                lease: None,
-                static_fee: None,
-                extra_fee: None,
-                max_fee: None,
-                validity_window: None,
-                first_valid_round: None,
-                last_valid_round: None,
-                account_references: None,
-                app_references: None,
-                asset_references: None,
-                box_references: Some(vec![BoxReference {
-                    app_id: 0,
-                    name: b"1".to_vec(),
-                }]),
-            },
-            None,
-        )
-        .await?;
-    match built {
-        algokit_transact::Transaction::AppCall(fields) => {
-            assert_eq!(fields.app_id, app_id);
-            assert_eq!(
-                fields.box_references.as_ref().unwrap(),
-                &vec![BoxReference {
-                    app_id: 0,
-                    name: b"1".to_vec()
-                }]
-            );
-        }
-        _ => panic!("expected app call"),
-    }
-    Ok(())
-}
-
-#[rstest]
-#[tokio::test]
-async fn construct_transaction_with_abi_encoding_including_transaction(
-    #[future] algorand_fixture: AlgorandFixtureResult,
-) -> TestResult {
-    let fixture = algorand_fixture.await?;
-    let sender = fixture.test_account.account().address();
-    // Use sandbox which has get_pay_txn_amount(pay)uint64
-    let spec = get_sandbox_spec();
-    let app_id = deploy_arc56_contract(&fixture, &sender, &spec, None, None).await?;
-
-    let mut algorand = RootAlgorandClient::default_localnet(None);
-    algorand.set_signer(sender.clone(), Arc::new(fixture.test_account.clone()));
-    let client = AppClient::new(AppClientParams {
-        app_id,
-        app_spec: spec,
-        algorand,
-        app_name: None,
-        default_sender: Some(sender.to_string()),
-        default_signer: None,
-        source_maps: None,
-        transaction_composer_config: None,
-    });
-
-    // Prepare a payment as an ABI transaction argument
-    let payment = PaymentParams {
-        sender: sender.clone(),
-        signer: None,
-        rekey_to: None,
-        note: None,
-        lease: None,
-        static_fee: None,
-        extra_fee: None,
-        max_fee: None,
-        validity_window: None,
-        first_valid_round: None,
-        last_valid_round: None,
-        receiver: sender.clone(),
-        amount: 12345,
-    };
-
-    let send_res = client
-        .send()
-        .call(
-            AppClientMethodCallParams {
-                method: "get_pay_txn_amount".to_string(),
-                args: vec![AppMethodCallArg::Payment(payment)],
-                sender: Some(sender.to_string()),
-                rekey_to: None,
-                note: None,
-                lease: None,
-                static_fee: None,
-                extra_fee: None,
-                max_fee: None,
-                validity_window: None,
-                first_valid_round: None,
-                last_valid_round: None,
-                account_references: None,
-                app_references: None,
-                asset_references: None,
-                box_references: None,
-            },
-            None,
-            None,
-        )
-        .await?;
-
-    // Expect a group of 2 transactions: payment + app call
-    assert_eq!(send_res.common_params.transactions.len(), 2);
-    // ABI return should be present and decode to expected value
-    let abi_ret = send_res.abi_return.as_ref().expect("abi return expected");
-    let ret_val = match &abi_ret.return_value {
-        Some(ABIValue::Uint(u)) => u.clone(),
-        _ => panic!("expected uint64 return"),
-    };
-    assert_eq!(ret_val, num_bigint::BigUint::from(12345u32));
-    Ok(())
-}
-
-#[rstest]
-#[tokio::test]
-async fn box_methods_with_arc4_returns_parametrized(
-    #[future] algorand_fixture: AlgorandFixtureResult,
-) -> TestResult {
-    let fixture = algorand_fixture.await?;
-    let sender = fixture.test_account.account().address();
-
-    // Deploy testing_app_puya
-    let spec =
-        Arc56Contract::from_json(algokit_test_artifacts::testing_app_puya::APPLICATION_ARC56)
-            .expect("valid arc56");
-    let app_id = deploy_arc56_contract(&fixture, &sender, &spec, None, None).await?;
-
-    let mut algorand = RootAlgorandClient::default_localnet(None);
-    algorand.set_signer(sender.clone(), Arc::new(fixture.test_account.clone()));
-    let client = AppClient::new(AppClientParams {
-        app_id,
-        app_spec: spec,
-        algorand,
-        app_name: None,
-        default_sender: Some(sender.to_string()),
-        default_signer: None,
-        source_maps: None,
-        transaction_composer_config: None,
-    });
-
-    // Fund app account to allow box writes
-    client
-        .fund_app_account(
-            FundAppAccountParams {
-                amount: 1_000_000,
-                sender: Some(sender.to_string()),
-                ..Default::default()
-            },
-            None,
-        )
-        .await?;
-
-    // Parametrized ARC-4 return cases
-    let mut big = num_bigint::BigUint::from(1u64);
-    big <<= 256u32;
-    let cases: Vec<(Vec<u8>, &str, &str, ABIValue)> = vec![
-        (
-            b"box_str".to_vec(),
-            "set_box_str",
-            "string",
-            ABIValue::from("string"),
-        ),
-        (
-            b"box_int".to_vec(),
-            "set_box_int",
-            "uint32",
-            ABIValue::from(123u32),
-        ),
-        (
-            b"box_int512".to_vec(),
-            "set_box_int512",
-            "uint512",
-            ABIValue::from(big),
-        ),
-        (
-            b"box_static".to_vec(),
-            "set_box_static",
-            "byte[4]",
-            ABIValue::Array(vec![
-                ABIValue::from_byte(1),
-                ABIValue::from_byte(2),
-                ABIValue::from_byte(3),
-                ABIValue::from_byte(4),
-            ]),
-        ),
-        // TODO: restore struct case after app factory is merged
-    ];
-
-    for (box_prefix, method_sig, value_type_str, arg_val) in cases {
-        // Encode the box name using ABIType "string"
-        let name_type = ABIType::from_str("string").unwrap();
-        let name_encoded = name_type.encode(&ABIValue::from("box1")).unwrap();
-        let mut box_reference = box_prefix.clone();
-        box_reference.extend_from_slice(&name_encoded);
-
-        // Send method call
-        client
-            .send()
-            .call(
-                AppClientMethodCallParams {
-                    method: method_sig.to_string(),
-                    args: vec![
-                        AppMethodCallArg::ABIValue(ABIValue::from("box1")),
-                        AppMethodCallArg::ABIValue(arg_val.clone()),
-                    ],
-                    sender: Some(sender.to_string()),
-                    rekey_to: None,
-                    note: None,
-                    lease: None,
-                    static_fee: None,
-                    extra_fee: None,
-                    max_fee: None,
-                    validity_window: None,
-                    first_valid_round: None,
-                    last_valid_round: None,
-                    account_references: None,
-                    app_references: None,
-                    asset_references: None,
-                    box_references: Some(vec![BoxReference {
-                        app_id: 0,
-                        name: box_reference.clone(),
-                    }]),
-                },
-                None,
-                None,
-            )
-            .await?;
-
-        // Verify raw equals ABI-encoded expected
-        let expected_raw = ABIType::from_str(value_type_str)
-            .unwrap()
-            .encode(&arg_val)
-            .unwrap();
-        let actual_raw = client.get_box_value(&box_reference).await?;
-        assert_eq!(actual_raw, expected_raw);
-
-        // Decode via ABI type and verify
-        let decoded = client
-            .algorand()
-            .app()
-            .get_box_value_from_abi_type(
-                client.app_id(),
-                &box_reference,
-                &ABIType::from_str(value_type_str).unwrap(),
-            )
-            .await?;
-        assert_eq!(decoded, arg_val);
-
-        // TODO: restore struct and nested struct tests after app factory is merged
-    }
-
-    Ok(())
-}
-
-#[rstest]
-#[tokio::test]
-async fn app_client_from_network_resolves_id(
-    #[future] algorand_fixture: AlgorandFixtureResult,
-) -> TestResult {
-    // Deploy hello_world and write networks mapping into spec, then call from_network
-    let fixture = algorand_fixture.await?;
-    let sender = fixture.test_account.account().address();
-
-    let spec = Arc56Contract::from_json(algokit_test_artifacts::hello_world::APPLICATION_ARC56)
-        .expect("valid arc56");
-    let app_id = deploy_arc56_contract(&fixture, &sender, &spec, None, None).await?;
-
-    let mut spec_with_networks = spec.clone();
-    spec_with_networks.networks = Some(std::collections::HashMap::from([(
-        "localnet".to_string(),
-        algokit_abi::arc56_contract::Network { app_id },
-    )]));
-
-    let client = AppClient::from_network(
-        spec_with_networks,
-        RootAlgorandClient::default_localnet(None),
-        None,
-        None,
-        None,
-        None,
-        None,
-    )
-    .await
-    .expect("from_network");
-    assert_eq!(client.app_id(), app_id);
-    Ok(())
-}
